Replaced repetitive code with loops
Hi- Saw your username on EngineerMan's recent pull requests.  It stuck out because I also am interested in mathematics.  This commit replaces 26 lines of code with three; it depends upon the sequential arrangements of characters in the ASCII table (type 'man ascii' at the terminal for a quick reference) and an implicit casting of int to char by printf.  Cheers